Add defaultapp redirect endpoint
