--- appinfo/app.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace OCA\AppOrder\AppInfo;
-
-\OCP\Util::addStyle('apporder', 'apporder');
-\OCP\Util::addScript('apporder', 'apporder');
-\OCP\App::registerAdmin('apporder', 'admin');
