--- appinfo/autoload.php
@@ -1,20 +1,0 @@
-<?php
-/**
- * ownCloud - foobar
- *
- * This file is licensed under the Affero General Public License version 3 or
- * later. See the COPYING file.
- *
- * @author jus <jus@bitgrid.net>
- * @copyright jus 2016
- */
-
-namespace OCA\AppOrder\AppInfo;
-use OCA\AppOrder\AppInfo\Application;
-
-use OCP\AppFramework\App;
-
-/**
- * Additional autoloader registration, e.g. registering composer autoloaders
- */
-// require_once __DIR__ . '/../vendor/autoload.php';
