+++ lib/AppInfo/Application.php
@@ -17,5 +17,8 @@ class Application extends App {
             );
         });
 
+		\OCP\Util::addStyle('apporder', 'apporder');
+		\OCP\Util::addScript('apporder', 'apporder');
+		\OCP\App::registerAdmin('apporder', 'admin');
     }
 }
